requirements.txt
nltk
psutil
psycopg2-binary
matplotlib